refactor: plus/cost 페이지 리팩터링#42 - price input 개선 - price clear button 추가 - 폴더 구조 개선
